feat: migrate NotificationController to NotificationService
