+++ client-app/lib/features/home/viewmodel/home_view_model.dart
@@ -1,20 +1,20 @@
 import 'package:flutter/material.dart';
 import 'package:plan_sync/controllers/app_preferences_controller.dart';
 import 'package:plan_sync/controllers/app_tour_controller.dart';
-import 'package:plan_sync/controllers/notification_controller.dart';
+import 'package:plan_sync/core/services/notification_service.dart';
 
 class HomeViewModel extends ChangeNotifier {
   HomeViewModel({
     required AppTourController appTour,
     required AppPreferencesController appPreferences,
-    required NotificationController notifications,
+    required NotificationService notifications,
   })  : _appTour = appTour,
         _appPreferences = appPreferences,
         _notifications = notifications;
 
   final AppTourController _appTour;
   final AppPreferencesController _appPreferences;
-  final NotificationController _notifications;
+  final NotificationService _notifications;
 
   GlobalKey get schedulePreferencesButtonKey =>
       _appTour.schedulePreferencesButtonKey;
