+++ client-app/lib/router_refresh_stream.dart
@@ -2,7 +2,7 @@ import 'dart:developer';
 import 'package:firebase_auth/firebase_auth.dart';
 import 'package:firebase_messaging/firebase_messaging.dart';
 import 'package:flutter/material.dart';
-import 'package:plan_sync/controllers/notification_controller.dart';
+import 'package:plan_sync/core/services/notification_service.dart';
 
 class GoRouterRefreshStream extends ChangeNotifier {
   GoRouterRefreshStream() {
@@ -13,7 +13,7 @@ class GoRouterRefreshStream extends ChangeNotifier {
       if (initialMessage != null) {
         // Handle navigation, e.g.:
         final route = initialMessage.data['route'];
-        NotificationController.initialNotificationRoute = route;
+        NotificationService.initialNotificationRoute = route;
         log('Route from initial message: $route');
         notifyListeners();
       }
